Updated PowerAvailable
Fixing issue identified during the pull request. The code has been updated to 'zero' the power available for any component that has a downstream connection, or is a propeller. The second condition is needed because propellers cannot provide any power on their own --- they must be powered by another component.